cleaning up loading function
