+++ autoarray/dataset/imaging/__init__.py
@@ -1,1 +1,3 @@
 from .imaging import *
+from .load import *
+from .masked_imaging import *
