+++ test_autoarray/unit/dataset/test_imaging.py
@@ -476,109 +476,6 @@ class TestEstimateDataGrid:
         assert sky_noise == np.std(np.arange(48))
 
 
-class TestNewImagingResized:
-    def test__all_components_resized__psf_is_not(self):
-        image = aa.array.manual_2d(array=np.ones((6, 6)), pixel_scales=1.0)
-        image[21] = 2.0
-
-        noise_map_array = aa.array.ones(shape_2d=(6, 6))
-        noise_map_array[21] = 3.0
-
-        background_noise_map_array = aa.array.ones(shape_2d=(6, 6))
-        background_noise_map_array[21] = 4.0
-
-        exposure_time_map_array = aa.array.ones(shape_2d=(6, 6))
-        exposure_time_map_array[21] = 5.0
-
-        background_sky_map_array = aa.array.ones(shape_2d=(6, 6))
-        background_sky_map_array[21] = 6.0
-
-        imaging = aa.imaging(
-            image=image,
-            psf=aa.kernel.zeros(shape_2d=(3, 3)),
-            noise_map=noise_map_array,
-            background_noise_map=background_noise_map_array,
-            exposure_time_map=exposure_time_map_array,
-            background_sky_map=background_sky_map_array,
-        )
-
-        imaging = imaging.resized_from_new_shape(new_shape=(4, 4))
-
-        assert (
-            imaging.image.in_2d
-            == np.array(
-                [
-                    [1.0, 1.0, 1.0, 1.0],
-                    [1.0, 1.0, 1.0, 1.0],
-                    [1.0, 1.0, 2.0, 1.0],
-                    [1.0, 1.0, 1.0, 1.0],
-                ]
-            )
-        ).all()
-        assert (
-            imaging.noise_map.in_2d
-            == np.array(
-                [
-                    [1.0, 1.0, 1.0, 1.0],
-                    [1.0, 1.0, 1.0, 1.0],
-                    [1.0, 1.0, 3.0, 1.0],
-                    [1.0, 1.0, 1.0, 1.0],
-                ]
-            )
-        ).all()
-        assert (
-            imaging.background_noise_map.in_2d
-            == np.array(
-                [
-                    [1.0, 1.0, 1.0, 1.0],
-                    [1.0, 1.0, 1.0, 1.0],
-                    [1.0, 1.0, 4.0, 1.0],
-                    [1.0, 1.0, 1.0, 1.0],
-                ]
-            )
-        ).all()
-        assert (
-            imaging.exposure_time_map.in_2d
-            == np.array(
-                [
-                    [1.0, 1.0, 1.0, 1.0],
-                    [1.0, 1.0, 1.0, 1.0],
-                    [1.0, 1.0, 5.0, 1.0],
-                    [1.0, 1.0, 1.0, 1.0],
-                ]
-            )
-        ).all()
-        assert (
-            imaging.background_sky_map.in_2d
-            == np.array(
-                [
-                    [1.0, 1.0, 1.0, 1.0],
-                    [1.0, 1.0, 1.0, 1.0],
-                    [1.0, 1.0, 6.0, 1.0],
-                    [1.0, 1.0, 1.0, 1.0],
-                ]
-            )
-        ).all()
-
-        assert imaging.poisson_noise_map == None
-
-        assert imaging.pixel_scales == (1.0, 1.0)
-        assert (imaging.psf.in_2d == np.zeros((3, 3))).all()
-        assert imaging.geometry.origin == (0.0, 0.0)
-
-    def test__resize_psf(self):
-        image = aa.array.manual_2d(array=np.ones((6, 6)))
-
-        imaging = aa.imaging(
-            image=image, noise_map=None, psf=aa.kernel.zeros(shape_2d=(3, 3))
-        )
-
-        imaging = imaging.resized_psf_from_new_shape(new_shape=(1, 1))
-
-        assert (imaging.image.in_2d == np.ones((6, 6))).all()
-        assert (imaging.psf.in_2d == np.zeros((1, 1))).all()
-
-
 class TestNewImagingModifiedImage:
     def test__imaging_returns_with_modified_image(self):
         image = aa.array.manual_2d(array=np.ones((4, 4)), pixel_scales=1.0)
